Test function to add a task to the store
